+++ src/migration/ember-app/steps/create-file-path-maps.ts
@@ -1,8 +1,8 @@
 import type { FilePathMap } from '@codemod-utils/files';
 
 import type { Options } from '../../../types/index.js';
-import { migrationStrategyForAppFolder } from './create-file-path-maps/app/index.js';
-import { migrationStrategyForTestsFolder } from './create-file-path-maps/tests/index.js';
+import { mapAppFolder } from './create-file-path-maps/app/index.js';
+import { mapTestsFolder } from './create-file-path-maps/tests/index.js';
 
 type MigrationStrategies = {
   app: FilePathMap;
@@ -10,11 +10,8 @@ type MigrationStrategies = {
 };
 
 export function createFilePathMaps(options: Options): MigrationStrategies {
-  const filePathMapApp = migrationStrategyForAppFolder(options);
-  const filePathMapTests = migrationStrategyForTestsFolder(options);
-
   return {
-    app: filePathMapApp,
-    tests: filePathMapTests,
+    app: mapAppFolder(options),
+    tests: mapTestsFolder(options),
   };
 }
