+++ src/migration/ember-app/steps/create-file-path-maps/app/map-route-controllers.ts
@@ -7,23 +7,24 @@ import type {
   Options,
 } from '../../../../../types/index.js';
 
-export function migrationStrategyForRouteAdapters(
-  options: Options,
-): FilePathMapEntries {
+export function mapRouteControllers(options: Options): FilePathMapEntries {
   const { podPath, projectRoot } = options;
 
-  const filePaths = findFiles(join('app', podPath, '**', 'adapter.{js,ts}'), {
-    projectRoot,
-  });
+  const filePaths = findFiles(
+    join('app', podPath, '**', 'controller.{js,ts}'),
+    {
+      projectRoot,
+    },
+  );
 
   return filePaths.map((oldFilePath) => {
     const newFilePath = renamePathByFile(oldFilePath, {
       find: {
         directory: join('app', podPath),
-        file: 'adapter',
+        file: 'controller',
       },
       replace: (key: string) => {
-        return `app/adapters/${key}`;
+        return `app/controllers/${key}`;
       },
     });
